# fixed bug and optimized axis_angle_to_rotation_matrix() by 1.5x