https://github.com/kornia/kornia/issues/2924 fixed the bug as described in this issue, also added tests to test the function and this bug specifically (the original code fails those tests), also optimized the function by 1.5x by avoiding recalculations and more efficient tensor ops

https://colab.research.google.com/drive/1wGiyX0oy9ySXpPlAjO5hlHIhJ4BWGdHL?usp=sharing

this is the benchmarking + validation script 

Device=cpu: Original=15.551 ms | Optimized=11.405 ms | Speedup=1.36×
Device=cuda: Original=1.456 ms | Optimized=0.968 ms | Speedup=1.50×