@cjpurackal 2025-09-04
Actually these feel a little redundant with the exponential map and `matrix()` of `kornia.geometry.liegroup.so3` https://github.com/kornia/kornia/blob/e779c63d54ac73bbaa13d11587d7a944125ebdd4/kornia/geometry/liegroup/so3.py#L110
https://github.com/kornia/kornia/blob/e779c63d54ac73bbaa13d11587d7a944125ebdd4/kornia/geometry/liegroup/so3.py#L211
Can you look into how these functions can be used in `SO3` ? Perhaps in a different PR.